+++ src/components/common/Loading/LoadingSpinner.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { CgSpinner } from 'react-icons/cg';
-import styles from './LoadingSpinner.module.css';
+import styles from './css/LoadingSpinner.module.css';
 
 export default function LoadingSpinner() {
   return (
