+++ src/components/common/Loading/css/LoadingDetail.module.css
