+++ src/pages/Login/Login.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Navigate } from 'react-router-dom';
 import styles from './Login.module.css';
-import LoadingMain from '../../components/common/LoadingMain/LoadingMain';
 import useLogin from './hooks/useLogin';
 import Symbole from '../../components/common/icons/Symbole';
+import LoadingMain from '../../components/common/Loading/LoadingMain';
 
 export default function Login() {
   const { isLoading, user, handleGoogleLogin, handleAnonymouseLogin, isWebView } = useLogin();
